tooltip is now much smarter
