+++ Core/Definitions.cs
@@ -390,12 +390,12 @@ namespace LiveCharts
         void SetDrawMarginLeft(double value);
         void SetDrawMarginHeight(double value);
         void SetDrawMarginWidth(double value);
-        void Erase();
+
         void AddToView(object element);
         void AddToDrawMargin(object element);
         void RemoveFromView(object element);
         void RemoveFromDrawMargin(object element);
-        void ShowTooltip(ChartPoint sender, IEnumerable<ChartPoint> sibilings, CorePoint at);
+
         void HideTooltop();
         void ShowLegend(CorePoint at);
         void HideLegend();
