+++ WpfView/DefaultTooltip.xaml
@@ -31,32 +31,61 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:wpf="clr-namespace:LiveCharts.Wpf"
              mc:Ignorable="d" 
-             BorderThickness="1" BorderBrush="#303030" d:DataContext="{d:DesignInstance wpf:DefaultTooltip}">
+             d:DataContext="{d:DesignInstance wpf:DefaultTooltip}"
+             BorderThickness="1.3">
+    <UserControl.Background>
+        <SolidColorBrush Color="#202020" Opacity=".8" />
+    </UserControl.Background>
     <UserControl.Resources>
         <Style TargetType="TextBlock">
             <Setter Property="Foreground" Value="{Binding Foreground}"></Setter>
         </Style>
+        <wpf:SharedConverter x:Key="SharedConverter"/>
+        <wpf:SharedVisibilityConverter x:Key="SharedVisibilityConverter"/>
+        <wpf:XVisibilityConverter x:Key="XVisibilityConverter"/>
+        <wpf:YVisibilityConverter x:Key="YVisibilityConverter"/>
+        <wpf:ParticipationVisibilityConverter x:Key="ParticipationVisibilityConverter"/>
     </UserControl.Resources>
-    <UserControl.Background>
-        <SolidColorBrush Color="#202020" Opacity=".9"></SolidColorBrush>
-    </UserControl.Background>
-    <Border CornerRadius="{Binding CornerRadius}">
-        <ItemsControl ItemsSource="{Binding Data}" Grid.IsSharedSizeScope="True">
-            <ItemsControl.ItemTemplate>
-                <DataTemplate DataType="{x:Type wpf:DataPointViewModel}">
-                    <Grid Margin="2">
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="Auto"/>
-                            <ColumnDefinition Width="Auto" SharedSizeGroup="Title"/>
-                            <ColumnDefinition Width="Auto" SharedSizeGroup="Value"/>
-                        </Grid.ColumnDefinitions>
-                        <Ellipse Width="11" Height="11" StrokeThickness="{Binding Series.StrokeThickness}" 
-                                 Stroke="{Binding Series.Stroke}" Fill="{Binding Series.Fill}"/>
-                        <TextBlock Grid.Column="1" Text="{Binding Series.Title}" Margin="4 0" VerticalAlignment="Center"/>
-                        <TextBlock Grid.Column="2" Text="{Binding Points}" VerticalAlignment="Center"/>
-                    </Grid>
-                </DataTemplate>
-            </ItemsControl.ItemTemplate>
-        </ItemsControl>
-    </Border>
+    <UserControl.Template>
+        <ControlTemplate>
+            <Border Background="{Binding Background}" CornerRadius="{Binding CornerRadius}"
+                    BorderThickness="{Binding BorderThickness}" Padding="10 5">
+                <StackPanel Orientation="Vertical">
+                    <TextBlock Text="{Binding Data, Converter={StaticResource SharedConverter}}" HorizontalAlignment="Center" FontWeight="Bold"
+                               Visibility="{Binding Data, Converter={StaticResource SharedVisibilityConverter}}" />
+                    <ItemsControl ItemsSource="{Binding Data.Points}" Grid.IsSharedSizeScope="True">
+                        <ItemsControl.ItemTemplate>
+                            <DataTemplate DataType="{x:Type wpf:DataPointViewModel}">
+                                <Grid Margin="2">
+                                    <Grid.ColumnDefinitions>
+                                        <ColumnDefinition Width="Auto"/>
+                                        <ColumnDefinition Width="Auto" SharedSizeGroup="Title"/>
+                                        <ColumnDefinition Width="Auto" SharedSizeGroup="X"/>
+                                        <ColumnDefinition Width="Auto" SharedSizeGroup="Y"/>
+                                        <ColumnDefinition Width="Auto" SharedSizeGroup="Participation"/>
+                                    </Grid.ColumnDefinitions>
+                                    <Ellipse Width="11" Height="11" StrokeThickness="{Binding Series.StrokeThickness}" 
+                                             Stroke="{Binding Series.Stroke}" Fill="{Binding Series.Fill}"/>
+                                    <TextBlock Grid.Column="1" Text="{Binding Series.Title}" VerticalAlignment="Center" Margin="5 0 0 0"/>
+                                    <TextBlock Grid.Column="2" Text="{Binding ChartPoint.X}" VerticalAlignment="Center" Margin="5 0 0 0"
+                                               Visibility="{Binding DataContext.Data, RelativeSource={RelativeSource  FindAncestor, 
+                                                                                                    AncestorType={x:Type StackPanel}}, 
+                                                                    Converter={StaticResource XVisibilityConverter} }" />
+                                    <TextBlock Grid.Column="3" Text="{Binding ChartPoint.Y}" VerticalAlignment="Center" Margin="5 0 0 0"
+                                               Visibility="{Binding DataContext.Data, RelativeSource={RelativeSource  FindAncestor, 
+                                                                                                    AncestorType={x:Type StackPanel}}, 
+                                                                    Converter={StaticResource YVisibilityConverter} }"/>
+                                    <TextBlock Grid.Column="4" Text="{Binding ChartPoint.Participation, StringFormat={}{0:P}}" 
+                                               VerticalAlignment="Center" Margin="5 0 0 0"
+                                               Visibility="{Binding DataContext.Data, RelativeSource={RelativeSource  FindAncestor, 
+                                                                                                    AncestorType={x:Type StackPanel}}, 
+                                                                    Converter={StaticResource ParticipationVisibilityConverter} }"/>
+                                </Grid>
+                            </DataTemplate>
+                        </ItemsControl.ItemTemplate>
+                    </ItemsControl>
+                </StackPanel>
+            </Border>
+        </ControlTemplate>
+    </UserControl.Template>
 </UserControl>
