+++ WpfView/Series/VerticalLineSeries.cs
@@ -30,6 +30,7 @@ using System.Windows.Media.Animation;
 using System.Windows.Shapes;
 using LiveCharts.SeriesAlgorithms;
 using LiveCharts.Wpf.Components;
+using LiveCharts.Wpf.Components.Chart;
 using LiveCharts.Wpf.Points;
 
 // ReSharper disable once CheckNamespace
